% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_train_test.R
\name{create_train_test}
\alias{create_train_test}
\title{Create Train/Test Datasets for Modeling}
\usage{
create_train_test(df, train_split)
}
\arguments{
\item{df}{dataframe to split for modeling}

\item{train_split}{float between 0.1 - 0.9, determines size of training data}
}
\value{
list containing two dataframes
}
\description{
Create Train/Test Datasets for Modeling
}
\examples{
data <- create_train_test(iris_df, 0.7)
training_data <- data$train
testing_data <- data$test
}
